Large Update - Check Commit Log For Details
Made a lot of new changes to multiple files and installed a new package.
Below is a short summary of changes made:

- Added UserPagesController.cs
- Added UserCaseModel.cs
- Added UserCases.cshtml
- Added Login.cshtml file
- Added Cases.cs

- Edited Program.cs

Installed "Pomelo.EntityFrameworkCore.MySql"

Created a new page for the user to view their submitted cases. Currently, this branch project does not include user data or connection to the database which means trying to access the user cases will result in an intentional error. This should be resolved once a user has been registered/made and the project is communicating properly with the database.